Rename ContentTypeNegotiation to NegotiateToResponse
Add case for a single type in NegotiateToResponse

ToService now use NegotiateToResponse in any case
